feat: QuotationProps 파일 분리, 결제 로직 주석 수정, 쿠폰 부분 주석처리(DURI-121)
