+++ apps/duri/src/pages/PaymentPage/index.tsx
@@ -1,22 +1,10 @@
 import { useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 
+import { QuotationProps } from '@duri/assets/types/QuotationType';
 import PaymentInfo from '@duri/components/payment/info';
 import PaymentWidget from '@duri/components/payment/widget';
 import { getQuotationInfo } from '@duri-fe/utils';
-
-export interface QuotationProps {
-  salonIdx: string;
-  salonName: string;
-  salonAddress: string;
-  salonImage: string;
-  designerName: string;
-  designerCareer: string;
-  designerAge: number;
-  designerGender: string;
-  groomingList: { menu: string; price: number }[]; // 배열의 각 요소가 "{menu:"시술 이름", price:"금액"}" 형식의 객체
-  groomingTotalPrice: number;
-}
 
 const PaymentPage = () => {
   const { quotationId } = useParams<{ quotationId: string }>();
@@ -39,13 +27,19 @@ const PaymentPage = () => {
     })();
   }, [quotationId]);
 
+  //쿠폰 정보 - 후순위!!!!!!
+  //   const [coupons, setCoupons] = useState();
+  useEffect(() => {
+    // 쿠폰 정보 불러와야 함
+  }, []);
+
   return (
     <>
       {quotationInfo ? (
         <>
-          {/* 결제 정보 컴포넌트 */}
-          <PaymentInfo quotationInfo={quotationInfo}/>
-          {/* 결제 방법 */}
+          {/* 결제 정보 컴포넌트 - 쿠폰 개발된다면 쿠폰 정보도 같이 넘겨야 함 */}
+          <PaymentInfo quotationInfo={quotationInfo} />
+          {/* 결제 방법 - 쿠폰 개발된다면 쿠폰 정보도 같이 넘겨야 함 */}
           <PaymentWidget quotationInfo={quotationInfo} />
         </>
       ) : (
